.maria/requirements.txt
torch
nltk
numpy
scikit-learn
datasets
streamlit
wandb
faiss-cpu
